Added partial class SongChart.AutoGeneration
